- merged proxy and adapter into only one file - only use DDP for Smarti Widget / Rocket.Chat messages (getConversation, getSmartiQueryBuilderResult) - added proxy endpoint for Smarti conversation search - added several comments
